refactor: Implement Money value object. Change entities that use BigDecimal for money concept to use Money.

+++ post-backend/parcel-service/src/main/java/com/jowk/parcel/catalog/dto/AdditionalServiceDetails.java
@@ -17,7 +17,7 @@ public record AdditionalServiceDetails(
         return new AdditionalServiceDetails(
                 additionalService.getId(),
                 additionalService.getName(),
-                additionalService.getPrice(),
+                additionalService.getPrice().toBigDecimal(),
                 additionalService.isAvailable()
         );
     }
